+++ application/core/MY_Controller.php
@@ -23,7 +23,6 @@ class Application extends CI_Controller
 
 		//  Set basic view parameters
 		$this->data = array ();
-		$this->data['pagetitle'] = 'CodeIgniter3.1 Starter 4';
 		$this->data['ci_version'] = (ENVIRONMENT === 'development') ? 'CodeIgniter Version <strong>'.CI_VERSION.'</strong>' : '';
 	}
 
@@ -32,7 +31,10 @@ class Application extends CI_Controller
 	 */
 	function render($template = 'template')
 	{
+	    $this->data['pagetitle'] = "Pizza Picker";
+        $this->data['menubar'] = $this->parser->parse('_menubar', $this->config->item('menu_choices'), true);
 		$this->data['content'] = $this->parser->parse($this->data['pagebody'], $this->data, true);
+		//$this->data['footer'] = $this->parser->parse('_footer', $this->config->item('footer'), true);
 		$this->parser->parse('template', $this->data);
 	}