+++ application/models/veg.php
@@ -1,0 +1,8 @@
+<?php
+
+class veg extends CSV_Model {
+    
+    public function __construct() {
+        parent::__construct('../data/veg.csv', 'ingrID');
+    }
+}