+++ application/views/view_catalog.php
@@ -1,0 +1,108 @@
+<div class="content center">
+    <h1>Pizza Topping List</h1>
+<table class="center">
+    <tr>
+        <th class="title">
+           Category
+        </th>
+        <th colspan="2">
+           Varieties
+        </th>
+
+    </tr>
+    <tr id="base">
+        <td class="title">
+            Bases
+        </td>
+        {bases}
+        <td>
+            <img src="{img}" style="width:150px;height:150px" title="{name}"/>
+            </br>
+            <b>{name}</b>
+            </br>
+            Type: {type}
+            </br>
+            price: {price}
+            </br>
+            Calories: {calories}
+        </td>
+        {/bases}
+    </tr>
+
+    <tr id="sauces">
+        <td class="title">
+            Sauces
+        </td>
+        {sauces}
+        <td>
+            <img src="{img}" style="width:150px;height:150px" title=""/>
+            </br>
+            <b>{name}</b>
+            </br>
+            Type: {type}
+            </br>
+            price: {price}
+            </br>
+            Calories: {calories}
+        </td>
+        {/sauces}
+    </tr>
+
+    <tr id="cheese">
+        <td class="title">
+            Cheese
+        </td>
+        {cheeses}
+        <td>
+            <img src="{img}" style="width:150px;height:150px" title=""/>
+            </br>
+            <b>{name}</b>
+            </br>
+            Type: {type}
+            </br>
+            price: {price}
+            </br>
+            Calories: {calories}
+        </td>
+        {/cheeses}
+    </tr>
+
+    <tr id="meat">
+        <td class="title">
+            Meats
+        </td>
+        {meats}
+        <td>
+            <img src="{img}" style="width:150px;height:150px" title=""/>
+            </br>
+            <b>{name}</b>
+            </br>
+            Type: {type}
+            </br>
+            price: {price}
+            </br>
+            Calories: {calories}
+        </td>
+        {/meats}
+    </tr>
+
+    <tr id="veg">
+        <td class="title">
+            Veggies
+        </td>
+        {veg}
+        <td>
+            <img src="{img}" style="width:150px;height:150px" title=""/>
+            </br>
+            <b>{name}</b>
+            </br>
+            Type: {type}
+            </br>
+            price: {price}
+            </br>
+            Calories: {calories}
+        </td>
+        {/veg}
+    </tr>
+    </table>
+</div>