+++ CHANGELOG.md
@@ -5,6 +5,49 @@ All notable changes to this project will be documented in this file.
 
 ## [Unreleased]
 
+## [0.0.7] - 2018-02-12
+- set - incomplete implementation of Set controller
+      - incomplete implementation of equipment_set view
+      - modified pizzas.csv to remove erroneous spacing
+
+## [0.0.6] - 2018-02-11
+### Changed
+- catalog - implemented catalog controller
+          - implemented view to list all ingredients by category
+          - modified ingredients.csv to remove erroneous spacing
+          - modified image path in ingredients.csv
+
+## [0.0.5] - 2018-02-11
+### Added
+- Base catalog view 
+- Catalog controller
+
+### Changed
+- routes - added separate route for catalog page
+- config - catalog goes to own page now
+
+## [0.0.4] - 2018-02-10
+### Added
+- Added csv 'database' files to /data, with corresponding models in application/models/
+### Changed
+- Modified image filenames and file extensions
+
+## [0.0.3] - 2018-02-09
+### Added
+- Added image files to /public/img/
+
+## [0.0.2] - 2018-02-08
+### Added
+- equipment set view
+- equipment set controller
+- menu
+### Changed
+- My_Controller - loads the menu
+- config - Adds the menu items
+- routes - Set home page to equipment_sets
+- template - added the menu as a navbar
+
+
 ## [0.0.1] - 2018-02-08
 ### Added
 - Changelog file